+++ src/App.vue
@@ -27,63 +27,12 @@
       </div>
     </template>
     <div class="container">
-      <section>
-        <div class="flex">
-          <div class="max-w-xs">
-            <label for="wallet" class="block text-sm font-medium text-gray-700"
-              >Тикер</label
-            >
-            <div class="mt-1 relative rounded-md shadow-md">
-              <input
-                @keyDown.enter="add"
-                v-model="ticker"
-                type="text"
-                name="wallet"
-                id="wallet"
-                class="block w-full pr-10 border-gray-300 text-gray-900 focus:outline-none focus:ring-gray-500 focus:border-gray-500 sm:text-sm rounded-md"
-                placeholder="Например DOGE"
-              />
-            </div>
-            <div
-              v-if="ticker !== '' && autocompletteTickers.length !== 0"
-              class="flex bg-white shadow-md p-1 rounded-md shadow-md flex-wrap"
-            >
-              <span
-                v-for="(item, index) in autocompletteTickers"
-                :key="index"
-                class="inline-flex items-center px-2 m-1 rounded-md text-xs font-medium bg-gray-300 text-gray-800 cursor-pointer"
-                @click="addFromAutocomplette(item)"
-              >
-                {{ item }}
-              </span>
-            </div>
-            <div v-if="tickerErrorMessage !== ''" class="text-sm text-red-600">
-              {{ tickerErrorMessage }}
-            </div>
-          </div>
-        </div>
-        <button
-          @click="add"
-          type="button"
-          class="my-4 inline-flex items-center py-2 px-4 border border-transparent shadow-sm text-sm leading-4 font-medium rounded-full text-white bg-gray-600 hover:bg-gray-700 transition-colors duration-300 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-gray-500"
-        >
-          <!-- Heroicon name: solid/mail -->
-          <svg
-            class="-ml-0.5 mr-2 h-6 w-6"
-            xmlns="http://www.w3.org/2000/svg"
-            width="30"
-            height="30"
-            viewBox="0 0 24 24"
-            fill="#ffffff"
-          >
-            <path
-              d="M13 7h-2v4H7v2h4v4h2v-4h4v-2h-4V7zm-1-5C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8s3.59-8 8-8 8 3.59 8 8-3.59 8-8 8z"
-            ></path>
-          </svg>
-          Добавить
-        </button>
-      </section>
-
+      <add-ticker-view
+        @add-ticker="add"
+        :disabled="tooManyTicketsAdded"
+        :tickerErrorMessage="tickerErrorMessage"
+        :coinsList="coinsList"
+      />
       <template v-if="tickers.length > 0">
         <hr class="w-full border-t border-gray-600 my-4" />
         <div>
@@ -149,50 +98,11 @@
         </dl>
         <hr class="w-full border-t border-gray-600 my-4" />
       </template>
-
-      <section class="relative" v-if="selectedTicker">
-        <h3 class="text-lg leading-6 font-medium text-gray-900 my-8">
-          VUE - USD
-        </h3>
-        <div
-          class="flex items-end border-gray-600 border-b border-l h-64"
-          ref="graph"
-        >
-          <div
-            v-for="(bar, index) in normalizedGraph"
-            :key="index"
-            :style="{ height: `${bar}%` }"
-            class="bg-purple-800 border w-10"
-          ></div>
-        </div>
-        <button
-          type="button"
-          class="absolute top-0 right-0"
-          @click="selectedTicker = null"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            xmlns:xlink="http://www.w3.org/1999/xlink"
-            xmlns:svgjs="http://svgjs.com/svgjs"
-            version="1.1"
-            width="30"
-            height="30"
-            x="0"
-            y="0"
-            viewBox="0 0 511.76 511.76"
-            style="enable-background:new 0 0 512 512"
-            xml:space="preserve"
-          >
-            <g>
-              <path
-                d="M436.896,74.869c-99.84-99.819-262.208-99.819-362.048,0c-99.797,99.819-99.797,262.229,0,362.048    c49.92,49.899,115.477,74.837,181.035,74.837s131.093-24.939,181.013-74.837C536.715,337.099,536.715,174.688,436.896,74.869z     M361.461,331.317c8.341,8.341,8.341,21.824,0,30.165c-4.16,4.16-9.621,6.251-15.083,6.251c-5.461,0-10.923-2.091-15.083-6.251    l-75.413-75.435l-75.392,75.413c-4.181,4.16-9.643,6.251-15.083,6.251c-5.461,0-10.923-2.091-15.083-6.251    c-8.341-8.341-8.341-21.845,0-30.165l75.392-75.413l-75.413-75.413c-8.341-8.341-8.341-21.845,0-30.165    c8.32-8.341,21.824-8.341,30.165,0l75.413,75.413l75.413-75.413c8.341-8.341,21.824-8.341,30.165,0    c8.341,8.32,8.341,21.824,0,30.165l-75.413,75.413L361.461,331.317z"
-                fill="#718096"
-                data-original="#000000"
-              ></path>
-            </g>
-          </svg>
-        </button>
-      </section>
+      <chart-view
+        @close="chartClose"
+        :selectedTicker="selectedTicker"
+        :tickers="tickers"
+      />
     </div>
   </div>
 </template>
@@ -204,28 +114,29 @@ import {
   unsubscribeFromTicker
 } from "./api.js";
 
+import AddTickerView from "./components/AddTickerView.vue";
+import ChartView from "./components/ChartView.vue";
+
 export default {
   name: "App",
+  components: {
+    AddTickerView,
+    ChartView
+  },
   data() {
     return {
-      ticker: "",
       tickers: [],
       selectedTicker: null,
-      graph: [],
       coinsList: [],
       loadingMask: true,
       tickerErrorMessage: "",
       filter: "",
-      page: 1,
-      maxGraphElements: 1
+      page: 1
     };
   },
   methods: {
-    calculateMaxGraphElements() {
-      if (!this.$refs) {
-        return;
-      }
-      this.maxGraphElements = this.$refs.graph.clientWidth / 38;
+    chartClose() {
+      this.selectedTicker = null;
     },
     formatPrice(price) {
       if (typeof price === "number") {
@@ -234,9 +145,9 @@ export default {
         return "-";
       }
     },
-    add() {
+    add(ticker) {
       const newTicker = {
-        name: this.ticker.toUpperCase(),
+        name: ticker,
         price: "-",
         invalid: false
       };
@@ -247,7 +158,7 @@ export default {
         this.tickerErrorMessage = "Already exists";
         return;
       }
-      if (this.ticker === "") {
+      if (ticker === "") {
         this.tickerErrorMessage = "Empty value";
         return;
       }
@@ -261,8 +172,6 @@ export default {
       }
 
       this.tickers = [...this.tickers, newTicker];
-
-      this.ticker = "";
       this.tickerErrorMessage = "";
       this.filter = "";
 
@@ -278,10 +187,6 @@ export default {
         })
       );
     },
-    addFromAutocomplette(item) {
-      this.ticker = item;
-      this.add();
-    },
     handleDelete(toRemove) {
       this.tickers = this.tickers.filter(t => t !== toRemove);
       if (this.selectedTicker === toRemove) {
@@ -289,10 +194,8 @@ export default {
       }
       unsubscribeFromTicker(toRemove.name);
     },
-
     select(t) {
       this.selectedTicker = t;
-      this.$nextTick().then(this.calculateMaxGraphElements);
     },
     updateTickers(tickerName, price, invalid = false) {
       this.tickers
@@ -300,18 +203,13 @@ export default {
         .forEach(t => {
           t.price = price;
           t.invalid = invalid;
-          if (t === this.selectedTicker) {
-            this.graph.push(price);
-            if (this.graph.length > this.maxGraphElements) {
-              this.graph = this.graph.slice(
-                this.graph.length - this.maxGraphElements
-              );
-            }
-          }
         });
     }
   },
   computed: {
+    tooManyTicketsAdded() {
+      return this.tickers.length > 4;
+    },
     startIndex() {
       return (this.page - 1) * 6;
     },
@@ -326,37 +224,12 @@ export default {
     paginatedTickers() {
       return this.filteredTickers.slice(this.startIndex, this.endIndex);
     },
-    autocompletteTickers() {
-      return this.coinsList
-        .filter(
-          item =>
-            item.symbol.includes(this.ticker.toLowerCase()) ||
-            item.fullname.includes(this.ticker.toLowerCase())
-        )
-        .slice(0, 4)
-        .map(item => item.symbol);
-    },
     hasNextPage() {
       return this.filteredTickers.length > this.endIndex;
-    },
-    normalizedGraph() {
-      const max = Math.max(...this.graph);
-      const min = Math.min(...this.graph);
-
-      const ret = this.graph.map(
-        price => 5 + (max - min !== 0 ? ((price - min) * 95) / (max - min) : 0)
-      );
-      return ret;
     },
     pageStateOptions() {
       return { filter: this.filter, page: this.page };
     }
-  },
-  mounted() {
-    window.addEventListener("resize", this.calculateMaxGraphElements);
-  },
-  beforeUnmount() {
-    window.removeEventListener("resize", this.calculateMaxGraphElements);
   },
   async created() {
     const data = await loadAllCoins();
@@ -400,9 +273,6 @@ export default {
     tickers() {
       localStorage.setItem("cryptonomicon-list", JSON.stringify(this.tickers));
     },
-    selectedTicker() {
-      this.graph = [];
-    },
     paginatedTickers() {
       if (this.paginatedTickers.length === 0 && this.page > 1) {
         this.page -= 1;
@@ -417,13 +287,6 @@ export default {
         document.title,
         `${window.location.pathname}?filter=${value.filter}&page=${value.page}`
       );
-    },
-    maxGraphElements() {
-      if (this.graph.length > this.maxGraphElements) {
-        this.graph = this.graph.slice(
-          this.graph.length - this.maxGraphElements
-        );
-      }
     }
   }
 };
