+++ src/components/AddButton.vue
@@ -1,0 +1,46 @@
+<template>
+  <button
+    :disabled="disabled"
+    type="button"
+    @click="click"
+    class="my-4 inline-flex items-center py-2 px-4 border border-transparent shadow-sm text-sm leading-4 font-medium rounded-full text-white bg-gray-600 hover:bg-gray-700 transition-colors duration-300 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-gray-500"
+    :class="{
+      'opacity-50': disabled
+    }"
+  >
+    <!-- Heroicon name: solid/mail -->
+    <svg
+      class="-ml-0.5 mr-2 h-6 w-6"
+      xmlns="http://www.w3.org/2000/svg"
+      width="30"
+      height="30"
+      viewBox="0 0 24 24"
+      fill="#ffffff"
+    >
+      <path
+        d="M13 7h-2v4H7v2h4v4h2v-4h4v-2h-4V7zm-1-5C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8s3.59-8 8-8 8 3.59 8 8-3.59 8-8 8z"
+      ></path>
+    </svg>
+    Добавить
+  </button>
+</template>
+
+<script>
+export default {
+  props: {
+    disabled: {
+      type: Boolean,
+      required: false,
+      default: false
+    }
+  },
+  emits: {
+    click: null
+  },
+  methods: {
+    click() {
+      this.$emit("click");
+    }
+  }
+};
+</script>
